+++ libs/@guardian/source-react-components/src/choice-card/ChoiceCard.stories.tsx
@@ -1,10 +1,7 @@
-import { ThemeProvider } from '@emotion/react';
-import { palette } from '@guardian/source-foundations';
 import type { Story } from '@storybook/react';
 import { SvgCamera } from '../../vendor/icons/SvgCamera';
 import { ChoiceCard } from './ChoiceCard';
 import type { ChoiceCardProps } from './ChoiceCard';
-import { choiceCardThemeDefault } from './theme';
 
 export default {
 	title: 'ChoiceCard',
@@ -37,60 +34,8 @@ export default {
 	},
 };
 
-const choiceCardThemeDark = {
-	textUnSelected: palette.neutral[86],
-	textSelected: palette.brand[400],
-	textHover: palette.brand[800],
-	textError: palette.error[500],
-	borderUnSelected: palette.neutral[86],
-	borderSelected: palette.brand[800],
-	borderHover: palette.brand[800],
-	borderError: palette.error[500],
-	backgroundUnSelected: palette.neutral[20],
-	backgroundHover: palette.neutral[20],
-	backgroundSelected: palette.neutral[100],
-	backgroundTick: palette.brand[500],
-};
-
 const Template: Story<ChoiceCardProps> = (args: ChoiceCardProps) => (
-	<div style={{ columns: 5 }}>
-		<ChoiceCard {...args} />
-
-		{/* Can use all available values */}
-		<ThemeProvider
-			theme={{
-				choiceCard: {
-					...choiceCardThemeDefault.choiceCard,
-					backgroundUnSelected: palette.brandAlt[200],
-					textLabel: palette.neutral[0],
-				},
-			}}
-		>
-			<ChoiceCard {...args} />
-		</ThemeProvider>
-
-		{/* Backwards compatible */}
-		<ThemeProvider theme={{ choiceCard: choiceCardThemeDark }}>
-			<ChoiceCard {...args} />
-		</ThemeProvider>
-
-		{/* Backwards compatible but not advised */}
-		<ThemeProvider theme={{ choiceCard: choiceCardThemeDark }}>
-			<ChoiceCard
-				{...args}
-				theme={{ backgroundUnSelected: palette.brandAlt[300] }}
-			/>
-		</ThemeProvider>
-
-		{/* New advised method with props */}
-		<ChoiceCard
-			{...args}
-			theme={{
-				...choiceCardThemeDark,
-				backgroundUnSelected: palette.brandAlt[300],
-			}}
-		/>
-	</div>
+	<ChoiceCard {...args} />
 );
 
 // *****************************************************************************
@@ -118,9 +63,4 @@ IconDefaultTheme.args = {
 	label: 'Camera',
 	// @ts-expect-error - Storybook maps 'JSX element' to <em>Option 1</em>
 	icon: 'JSX element',
-};
-
-export const DarkTheme = Template.bind({});
-DarkTheme.args = {
-	theme: choiceCardThemeDark,
 };
